Class to own file. Xml Comments. public to internal.
