+++ src/HdrHistogram/HdrHistogram.csproj
@@ -57,6 +57,10 @@
     <Compile Include="HistogramExtensions.cs" />
     <Compile Include="HistogramLogReader.cs" />
     <Compile Include="HistogramLogWriter.cs" />
+    <Compile Include="Iteration\AllValueEnumerable.cs" />
+    <Compile Include="Iteration\LinearBucketEnumerable.cs" />
+    <Compile Include="Iteration\LogarithmicBucketEnumerable.cs" />
+    <Compile Include="Iteration\PercentileEnumerable.cs" />
     <Compile Include="OutputScalingFactor.cs" />
     <Compile Include="Persistence\IntCountsDecoder.cs" />
     <Compile Include="Persistence\LongCountsDecoder.cs" />
