+++ src/HdrHistogram/Iteration/LinearEnumerator.cs
@@ -8,43 +8,15 @@
  * https://github.com/HdrHistogram/HdrHistogram
  */
 
-using System.Collections;
-using System.Collections.Generic;
-
 namespace HdrHistogram.Iteration
 {
-    /// <summary>
-    /// An enumerator of <see cref="HistogramIterationValue"/> through the histogram using a <see cref="LinearEnumerator"/>
-    /// </summary>
-    sealed class LinearBucketEnumerable : IEnumerable<HistogramIterationValue>
-    {
-        private readonly HistogramBase _histogram;
-        private readonly int _valueUnitsPerBucket;
-
-        public LinearBucketEnumerable(HistogramBase histogram, int valueUnitsPerBucket)
-        {
-            this._histogram = histogram;
-            this._valueUnitsPerBucket = valueUnitsPerBucket;
-        }
-
-        public IEnumerator<HistogramIterationValue> GetEnumerator()
-        {
-            return new LinearEnumerator(_histogram, _valueUnitsPerBucket);
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return this.GetEnumerator();
-        }
-    }
-
     /// <summary>
     /// Used for iterating through histogram values in linear steps. The iteration is
     /// performed in steps of<i>valueUnitsPerBucket</i> in size, terminating when all recorded histogram
     /// values are exhausted. Note that each iteration "bucket" includes values up to and including
     /// the next bucket boundary value.
     /// </summary>
-    sealed class LinearEnumerator : AbstractHistogramEnumerator
+    internal sealed class LinearEnumerator : AbstractHistogramEnumerator
     {
         private readonly long _valueUnitsPerBucket;
         private long _nextValueReportingLevel;
